feat(tailwind-plugins): export TCV utils
